pincorder/requirements.txt
markdown
django
djangorestframework
django-oauth-toolkit